requirements.txt
django==3.1.7
django-allauth
djangorestframework==3.12.2
django-rest-auth==0.9.1
django-cors-headers==3.7.0
drf-yasg==1.20
eth-account==0.5.3
eth-utils==1.10.0
ethereum==2.0.4
gunicorn==20.0.4
pika==1.2.0
pillow==8.1.1
psycopg2-binary==2.8.6
rlp==0.6.0
web3==5.17.0
ipfshttpclient==0.7.0
httpcore
httpx
